CHG: Move virtual frame parsing to Level1Stripmap class.
